requirements.txt
fastapi
uvicorn
streamlit
requests
sentence-transformers
faiss-cpu
numpy
